Move some large static things to ROM Made Ssinonlytable[] and gamma_correction_table[] 32bit for aligned reads from ROM

+++ firmware/embeddedcommon/DFT32.h
@@ -3,7 +3,7 @@
 #ifndef _DFT32_H
 #define _DFT32_H
 
-extern const int16_t Ssinonlytable[];
+extern const int32_t Ssinonlytable[];
 
 #endif
